move typedef tests to new test
